+++ Controls/_toolbars/View.ts
@@ -2,11 +2,15 @@ import {ICrudPlus, PrefetchProxy} from 'Types/source';
 import {SyntheticEvent} from 'Vdom/Vdom';
 import {factory, RecordSet} from 'Types/collection';
 import {descriptor, Record} from 'Types/entity';
-import {ActualApi} from 'Controls/buttons';
 
 import {Control, IControlOptions, TemplateFunction} from 'UI/Base';
 import {Sticky  as StickyOpener} from 'Controls/popup';
-import {IShowType, showType, getMenuItems, needShowMenu} from 'Controls/Utils/Toolbar';
+import {showType, getMenuItems, needShowMenu} from 'Controls/Utils/Toolbar';
+
+import {
+    getButtonTemplate, hasSourceChanged,
+    getButtonTemplateOptionsByItem, getTemplateByItem, loadItems
+} from 'Controls/_toolbars/Util';
 import {IStickyPopupOptions, IStickyPosition, IEventHandlers} from 'Controls/popup';
 
 import {
@@ -26,53 +30,11 @@ import {IGrouped, IGroupedOptions} from 'Controls/dropdown';
 import * as template from 'wml!Controls/_toolbars/View';
 import * as defaultItemTemplate from 'wml!Controls/_toolbars/ItemTemplate';
 import * as ActualAPI from 'Controls/_toolbars/ActualAPI';
-import {ButtonTemplate, cssStyleGeneration} from 'Controls/buttons';
-import {CrudWrapper} from 'Controls/dataSource';
 
 type TItem = Record;
 type TItems = RecordSet<TItem>;
 type TypeItem = 'toolButton' | 'icon' | 'link' | 'list';
 export type TItemsSpacing = 'medium' | 'big';
-
-export function getButtonTemplateOptionsByItem(item: TItem, toolbarOptions: IControlOptions = {}): IButtonOptions {
-    const icon = item.get('icon');
-    const style = item.get('buttonStyle');
-    const viewMode = item.get('viewMode');
-
-    // todo: https://online.sbis.ru/opendoc.html?guid=244a5058-47c1-4896-a494-318ba2422497
-    const inlineHeight = viewMode === 'functionalButton' ? 'default' : ActualApi.actualHeight('m', undefined, viewMode, false);
-    const iconSize = viewMode === 'functionalButton' ? 's' : 'm';
-
-    const iconStyle = item.get('iconStyle');
-    const transparent = item.get('buttonTransparent');
-    const caption = item.get('caption');
-    const captionPosition = item.get('captionPosition');
-    const readOnly = item.get('readOnly') || toolbarOptions.readOnly;
-    const fontColorStyle = item.get('fontColorStyle');
-    const contrastBackground = item.get('contrastBackground');
-    const cfg: IButtonOptions = {};
-    cfg._hoverIcon = true;
-    cssStyleGeneration.call(cfg, {
-        inlineHeight,
-        icon,
-        style,
-        viewMode,
-        iconStyle,
-        iconSize,
-        transparent,
-        caption,
-        captionPosition,
-        readOnly,
-        fontColorStyle,
-        contrastBackground
-    });
-    cfg.readOnly = readOnly;
-    return cfg;
-}
-
-export function getButtonTemplate(): TemplateFunction {
-    return ButtonTemplate;
-}
 
 // Перейти на интерфейс выпадающих списков, когда он появится
 
@@ -150,24 +112,24 @@ export interface IToolbarOptions extends IControlOptions, IHierarchyOptions, IIc
  * @author Красильников А.С.
  * @demo Controls-demo/Toolbar/Base/Index
  */
-class Toolbar extends Control<IToolbarOptions, TItems> implements IHierarchy, IIconSize, IItemTemplate, IGrouped, IToolbarSource {
+class Toolbar extends Control<IToolbarOptions, TItems> implements IHierarchy, IIconSize, IItemTemplate,
+    IGrouped, IToolbarSource {
     /*
      * Used in template
      */
-    protected _showType: IShowType = showType;
     protected _needShowMenu: boolean = null;
     protected _fullItemsList: TItems = null;
     protected _items: TItems = null;
     protected _menuItems: TItems = null;
     protected _source: ICrudPlus = null;
-    protected _originalSource = null;
+    protected _originalSource: ICrudPlus = null;
     protected _menuSource: ICrudPlus = null;
     protected _nodeProperty: string = null;
     protected _parentProperty: string = null;
     protected _menuOptions: object = null;
     protected _isLoadMenuItems: boolean = false;
     protected _buttonTemplate: TemplateFunction = getButtonTemplate();
-    protected _actualItems = null;
+    protected _actualItems: TItems = null;
 
     protected _template: TemplateFunction = template;
 
@@ -317,14 +279,6 @@ class Toolbar extends Control<IToolbarOptions, TItems> implements IHierarchy, II
         this._needShowMenu = needShowMenu(this._actualItems);
     }
 
-    private _setStateBySource(source: ICrudPlus): Promise<TItems> {
-        return Toolbar._loadItems(source).then((items) => {
-            this._setStateByItems(items, source);
-
-            return items;
-        });
-    }
-
     private _needChangeState(newOptions: IToolbarOptions): boolean {
         const currentOptions = this._options;
 
@@ -336,14 +290,8 @@ class Toolbar extends Control<IToolbarOptions, TItems> implements IHierarchy, II
         ].some((optionName: string) => currentOptions[optionName] !== newOptions[optionName]);
     }
 
-    private _hasSourceChanged(newSource?: ICrudPlus) {
-        const currentSource = this._options.source;
-
-        return newSource && currentSource !== newSource;
-    }
-
-    private _openMenu(config): void {
-        StickyOpener.openPopup(config, this).then((popupId) => {
+    private _openMenu(config: IMenuOptions): void {
+        StickyOpener.openPopup(config).then((popupId) => {
             this._popupId = popupId;
         });
     }
@@ -355,7 +303,7 @@ class Toolbar extends Control<IToolbarOptions, TItems> implements IHierarchy, II
         if (receivedItems) {
             this._setStateByItems(receivedItems, options.source);
         } else if (options.source) {
-            return this._setStateBySource(options.source);
+            return this.setStateBySource(options.source);
         }
     }
 
@@ -363,10 +311,10 @@ class Toolbar extends Control<IToolbarOptions, TItems> implements IHierarchy, II
         if (this._needChangeState(newOptions)) {
             this._setState(newOptions);
         }
-        if (this._hasSourceChanged(newOptions.source)) {
+        if (hasSourceChanged(newOptions.source, this._options.source)) {
             this._originalSource = newOptions.source;
             this._isLoadMenuItems = false;
-            this._setStateBySource(newOptions.source);
+            this.setStateBySource(newOptions.source);
         }
     }
 
@@ -382,6 +330,13 @@ class Toolbar extends Control<IToolbarOptions, TItems> implements IHierarchy, II
                 StickyOpener.closePopup(this._popupId);
             }
         }
+    }
+
+    protected setStateBySource(source: ICrudPlus): Promise<TItems> {
+        return loadItems(source).then((items) => {
+            this._setStateByItems(items, source);
+            return items;
+        });
     }
 
     protected _closeHandler(): void {
@@ -415,13 +370,7 @@ class Toolbar extends Control<IToolbarOptions, TItems> implements IHierarchy, II
     }
 
     protected _getTemplateByItem(item: TItem): TemplateFunction {
-        const selfItemTemplate: TemplateFunction = item.get(this._options.itemTemplateProperty);
-
-        if (selfItemTemplate) {
-            return selfItemTemplate;
-        }
-
-        return this._options.itemTemplate;
+        return getTemplateByItem(item, this._options);
     }
 
     protected _getButtonTemplateOptionsByItem(item: TItem): IButtonOptions {
@@ -484,12 +433,6 @@ class Toolbar extends Control<IToolbarOptions, TItems> implements IHierarchy, II
         }
 
         return '';
-    }
-
-    private static _loadItems(source: ICrudPlus): Promise<TItems> {
-        const crudWrapper = new CrudWrapper({source});
-
-        return crudWrapper.query({});
     }
 
     private static _calcMenuItems(items: TItems): TItems {
