+++ Controls/toolbars.ts
@@ -41,5 +41,6 @@ import ItemTemplate = require('wml!Controls/_toolbars/ItemTemplate');
 export {ItemTemplate};
 
 export {items as actualItems} from 'Controls/_toolbars/ActualAPI';
-export {default as View, IToolbarOptions, TItemsSpacing, getButtonTemplateOptionsByItem, getButtonTemplate} from './_toolbars/View';
+export {default as View, IToolbarOptions, TItemsSpacing} from './_toolbars/View';
+export {getButtonTemplateOptionsByItem, getButtonTemplate} from './_toolbars/Util';
 export {default as BoxView} from './_toolbars/BoxView';
